+++ core/api/streaming_json_route.py
@@ -88,6 +88,7 @@ def _streaming_json_route[ApiRequest: BaseModel, ApiResponse: BaseModel](
                 validationErrorMessage = ', '.join([f'{".".join([str(value) for value in error["loc"]])}: {error["msg"]}' for error in exception.errors()])
                 raise BadRequestException(f'Invalid request: {validationErrorMessage}')
             kibaRequest: KibaApiRequest[ApiRequest] = KibaApiRequest(scope=receivedRequest.scope, receive=receivedRequest._receive, send=receivedRequest._send)  # noqa: SLF001
+            kibaRequest.originIp = typing.cast('str | None', receivedRequest.scope.get('originIp'))
             kibaRequest.data = requestParams
             responseGeneratorOrAwaitable = func(kibaRequest)
             responseGenerator = await responseGeneratorOrAwaitable if inspect.isawaitable(responseGeneratorOrAwaitable) else responseGeneratorOrAwaitable